Add autoshutdown script
Similar to autopuase, autoshutdown will shutdown the container after a
period of inactivity of java client connections.
The autoshutdown script reuses a lot of the same code as autopause with
the exception of knocking on ports to resume the server. After a period
of inactivity the script will kill the main java process causing the
container to stop.